feat: v26.41.4 巨石文件拆分 + UI优化 + 打开文件夹按钮禁用
- court_zxfw_guarantee.py (2741行) 拆分为 guarantee/ 子目录 6 个模块
- executor.py (2496行) 拆分为 6 个 mixin + 门面类
- "打开文件夹"按钮无绑定时自动禁用
- 案件案号区域 UI 精致化排版
- 裁判文书拖拽上传、LLM 智能模型路由等已在前序提交中完成
- ratchet 基线更新 + CHANGELOG/README 版本号更新
