+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 本项目的所有重要更改都将记录在此文件中。
 
+## [26.41.4] - 2026-04-30
+
+### 新功能
+
+- **LLM 智能模型路由**：新增模型路由器，按任务类型（抽取/评分/生成）自动选择最优模型；诉讼费计算、法律研究、合同审查三个模块接入 Qwen 397B
+- **裁判文书文件拖拽上传**：案件案号区域支持拖拽上传裁判文书 PDF，替代旧的文件选择按钮，使用 Alpine.js + Dropzone 实现
+- **LLM 默认后端动态配置**：`LLM_DEFAULT_BACKEND` 支持从 SystemConfig 数据库动态读取，无需重启即可切换默认 LLM 后端
+
+### 重构
+
+- **`court_zxfw_guarantee.py` 巨石文件拆分**：2741 行单文件拆分为 `guarantee/` 子目录 6 个模块（`base_mixin`、`data_mixin`、`dialog_mixin`、`form_filling_mixin`、`upload_mixin`、`guarantee_service`），按职责分离基础操作、数据处理、对话框、表单填充、文件上传
+- **`executor.py` 巨石文件拆分**：法律检索执行器从单文件拆分为 6 个 mixin（`intent_mixin`、`query_mixin`、`scoring_mixin`、`feedback_mixin`、`cache_mixin`、`policy_mixin`）+ 门面类，通过 MRO 组合继承
+- **案件案号区域 UI 精致化**：重新设计案号区字段布局，标签在上/输入框在下，复选框同行显示，fieldset 合并减少分组，去掉多余分割线和边框，按钮样式现代化
+- **移除旧上传 UI**：案件文书上传完全由 Dropzone 拖拽组件替代，移除旧的文件选择按钮代码
+
+### 修复
+
+- **裁判文书解析判决主文丢失**：`MAX_TEXT_LIMIT` 截断导致判决主文被丢弃，调整截断策略保留关键段落
+- **合并 PDF 排版句子内换行**：修复 PDF 合并过程中因排版导致的句子内异常换行
+- **案号带后缀提取**：支持案号带 `-1`、`-2` 等后缀的正确提取
+- **"打开文件夹"按钮无绑定时禁用**：案件案号区的"打开文件夹"按钮在案件本身或关联合同均未绑定文件夹时自动禁用，提示"未绑定文件夹，请先在下方绑定"
+
 ## [26.41.3] - 2026-04-30
 
 ### 修复
