+++ README.md
@@ -1,4 +1,4 @@
-# 法穿 AI 案件管理系统 · V26.41.3
+# 法穿 AI 案件管理系统 · V26.41.4
 
 *由一线执业律师主导研发 · 开源 · 私有化部署 · 数据自持*
