+++ backend/apps/automation/services/scraper/sites/guarantee/__init__.py
@@ -1,0 +1,5 @@
+"""一张网申请担保服务 — mixin 拆分包。"""
+
+from .guarantee_service import CourtZxfwGuaranteeService
+
+__all__ = ["CourtZxfwGuaranteeService"]
