+++ backend/apps/legal_research/services/executor_components/__init__.py
@@ -1,9 +1,21 @@
+from .cache_mixin import ExecutorCacheMixin
+from .feedback_mixin import ExecutorFeedbackMixin
+from .intent_mixin import ExecutorIntentMixin
+from .policy_mixin import ExecutorPolicyMixin
+from .query_mixin import ExecutorQueryMixin
 from .result_persistence import ExecutorResultPersistenceMixin
+from .scoring_mixin import ExecutorScoringMixin
 from .source_gateway import ExecutorSourceGatewayMixin
 from .task_lifecycle import ExecutorTaskLifecycleMixin
 
 __all__ = [
+    "ExecutorCacheMixin",
+    "ExecutorFeedbackMixin",
+    "ExecutorIntentMixin",
+    "ExecutorPolicyMixin",
+    "ExecutorQueryMixin",
     "ExecutorResultPersistenceMixin",
+    "ExecutorScoringMixin",
     "ExecutorSourceGatewayMixin",
     "ExecutorTaskLifecycleMixin",
 ]
